added and updated

+++ tests/test_api/test_routes/user_router_test.py
@@ -24,10 +24,10 @@ def test_current_user_sign():
     assert response.status_code == 200
 
 
-def test_select_last_user_unsign():
+def test_select_user_unsign():
     with client as init_client:
         response = init_client.get("/api/v1/user/9223372036854775807")
-    assert response.status_code == 404
+    assert response.status_code == 403
 
 
 def test_unsign_update_user():
@@ -58,20 +58,33 @@ def test_update_user_sign():
 
 def test_delete_user_unsign():
     with client as init_client:
-        response = init_client.delete("/api/v1/user/delete")
+        response = init_client.delete("/api/v1/user/delete", params={"user_id": 1})
     assert response.status_code == 403
 
 
 def test_delete_user_sign():
     # test delete user sign
     with client as init_client:
+        admin_data = get_register_data(init_client)
         user_data = register_user(init_client)
-        response = init_client.delete(
-            "/api/v1/user/delete",
+
+        # get user id
+        user_id = init_client.get(
+            "/api/v1/user/current",
             cookies=dict(
                 session_id=user_data["session_id"],
                 access_token=user_data["access_token"],
                 refresh_token=user_data["refresh_token"]
+            )
+        ).json().get("id")
+
+        response = init_client.delete(
+            "/api/v1/user/delete",
+            params={"user_id": user_id},
+            cookies=dict(
+                session_id=admin_data["session_id"],
+                access_token=admin_data["access_token"],
+                refresh_token=admin_data["refresh_token"]
             )
         )
     assert response.status_code == 204
